\name{acceptor.m}
\alias{acceptor.m}
\docType{data}
\title{
acceptor.m
}
\description{
Acceptor site mammalian frequency matrices for GT-AG pairs from SpliceDB
}
\usage{data("acceptor.m")}
\format{
  The format is:
 num [1:4, 1:15] 9 31.03 12.5 42.36 8.44 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:4] "A" "C" "G" "T"
  ..$ : chr [1:15] "V1" "V2" "V3" "V4" ...
}
\source{
\url{http://www.softberry.com/spldb/SpliceDB.html}
}
\references{
Burset M., Seledtsov I., Solovyev V. (Nucl.Acids Res.,2000,28,4364-4375; Nucl. Acids Res.,2001,29,255-259)
}
\examples{
data(acceptor.m)
}
\keyword{datasets}
