% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_splinter.R
\name{addEnsemblAnnotation}
\alias{addEnsemblAnnotation}
\title{addEnsemblAnnotation}
\usage{
addEnsemblAnnotation(data, species = "hsapiens")
}
\arguments{
\item{data}{\code{\link{extractSpliceEvents}} object}

\item{species}{character. biomaRt species passed to retrieve annotation.
Common species include: 'hsapiens','mmusculus'}
}
\value{
\code{\link{extractSpliceEvents}} object with annotated genes under $geneSymbol
}
\description{
Adds annotation to \code{\link{extractSpliceEvents}} object (if not present)
}
\examples{
data_path<-system.file("extdata",package="SPLINTER")
splice_data<-extractSpliceEvents(data=paste(data_path,"/skipped_exons.txt",sep=""))
splice_data<-addEnsemblAnnotation(data=splice_data,species="mmusculus")
}
\author{
Diana Low
}
\seealso{
\url{http://asia.ensembl.org/info/data/biomart/biomart_r_package.html#biomartexamples}
}

