% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_splinter.R
\name{annotateEvents}
\alias{annotateEvents}
\title{annotateEvents}
\usage{
annotateEvents(
  thedata,
  db,
  bsgenome,
  outputdir,
  full_output = FALSE,
  output_prefix = "results"
)
}
\arguments{
\item{thedata}{list. output of extractSpliceEvents.}

\item{db}{TxDb object}

\item{bsgenome}{BSGenome object}

\item{outputdir}{character. relative output directory to current location.}

\item{full_output}{logical. writes out detailed text report and generate figures.}

\item{output_prefix}{character. text prefix for full_output files.}
}
\value{
list containing information on
(1) data.frame with splicing regions
(2) splice event type
}
\description{
Gives detailed description of splicing event in terms of splicing outcome
post translation. Currently supports exon skipping and intron retention events.
}
\author{
Diana LOW
}
