% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primerpcr.R
\name{checkPrimer}
\alias{checkPrimer}
\title{checkPrimer}
\usage{
checkPrimer(pp, genome, roi = NULL)
}
\arguments{
\item{pp}{data.frame defining primers, or output of \code{\link{callPrimer3}}.
minimal columns = PRIMER_LEFT_SEQUENCE,PRIMER_RIGHT_SEQUENCE}

\item{genome}{BSgenome object}

\item{roi}{\code{\link{makeROI}} object}
}
\value{
list of GRanges with primer locations
}
\description{
checkPrimer
}
\examples{
# create a primer pair
roi
primer_pair <- data.frame(PRIMER_LEFT_SEQUENCE="agctcttgaaattggagctgac",
                          PRIMER_RIGHT_SEQUENCE="cttagaaagaacaggaaatcc",
                          stringsAsFactors=FALSE)
}
\author{
Diana Low
}
