\name{donor.m}
\alias{donor.m}
\docType{data}
\title{
donor.m
}
\description{
Donor site mammalian frequency matrices for GT-AG pairs from SpliceDB
}
\usage{data("donor.m")}
\format{
  The format is:
 num [1:4, 1:9] 34.1 36.2 18.3 11.4 60.4 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:4] "A" "C" "G" "T"
  ..$ : chr [1:9] "V1" "V2" "V3" "V4" ...
}
\source{
\url{http://www.softberry.com/spldb/SpliceDB.html}
}
\references{
Burset M., Seledtsov I., Solovyev V. (Nucl.Acids Res.,2000,28,4364-4375; Nucl. Acids Res.,2001,29,255-259)
}
\examples{
data(donor.m)
}
\keyword{datasets}
