% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_splinter.R
\name{eventOutcomeTranslate}
\alias{eventOutcomeTranslate}
\title{eventOutcomeTranslate}
\usage{
eventOutcomeTranslate(
  seq1,
  genome,
  direction = FALSE,
  fullseq = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{seq1}{GRangesList}

\item{genome}{BSGenome object}

\item{direction}{logical. Report direction of sequence change.}

\item{fullseq}{logical. Output full AA sequence.}

\item{verbose}{logical. turn messages on/off.}
}
\value{
list of translated sequences
}
\description{
translates sequences, reports if NMD or NTC
}
\examples{
suppressMessages(library(BSgenome.Mmusculus.UCSC.mm9))
bsgenome<-BSgenome.Mmusculus.UCSC.mm9
translation_results<-eventOutcomeTranslate(compatible_cds,genome=bsgenome,
 direction=TRUE)
}
\author{
Diana LOW
}
