% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_splinter.R
\name{extendROI}
\alias{extendROI}
\title{extendROI}
\usage{
extendROI(roi, tx, up = 0, down = 0, type = 1)
}
\arguments{
\item{roi}{\code{\link{makeROI}} object}

\item{tx}{GRangesList transcript list to pull regions from}

\item{up}{integer. number of exons to extend upstream}

\item{down}{integer. number of exons to extend downstream}

\item{type}{integer. 1=full cassette, 2=flank only}
}
\value{
\code{\link{makeROI}} object with modified ranges
}
\description{
extend the span of the current ROI by n number of up/downstream exon(s)
by modifying roi_range within the makeROI object while
retaining legacy sites by keeping $roi and $flank
}
\examples{
extendROI(roi,valid_tx,up=1)
}
