% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_splinter.R
\name{findCompatibleEvents}
\alias{findCompatibleEvents}
\title{findCompatibleEvents}
\usage{
findCompatibleEvents(tx, tx2 = NULL, roi, sequential = TRUE, verbose = FALSE)
}
\arguments{
\item{tx}{GRangesList object of transcripts}

\item{tx2}{optional GRangesList object of transcripts if tx is list of cds}

\item{roi}{\code{\link{makeROI}} object contaning event information}

\item{sequential}{logical. Exons have to appear sequentially to
be considered compatible}

\item{verbose}{logical. printouts and messages.}
}
\value{
list of length 4\cr
 (1) GRangesList\cr
 (2) Hits status [c]=coding; [nc]=non-coding\cr
 (3) ct - compatible transcripts\cr
 (4) tt - total transcripts
}
\description{
Which transcript contains the event?
Each event has 2 possibilities, as long as the transcript fulfills one,
it passes the test
Has to be exact (inner junctions)
}
\details{
Seperates into event/region1 and 2 for the alternative case
}
\examples{
compatible_cds <- findCompatibleEvents(valid_cds,roi=roi,verbose=TRUE)
}
\author{
Diana Low
}
