% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_splinter.R
\name{findCompatibleExon}
\alias{findCompatibleExon}
\title{findCompatibleExon}
\usage{
findCompatibleExon(tx, roi, verbose = FALSE)
}
\arguments{
\item{tx}{GRangesList object of transcripts}

\item{roi}{\code{\link{makeROI}} object contaning event information}

\item{verbose}{logical. printouts and messages.}
}
\value{
list of length 3\cr
 (1) GRangesList hits\cr
 (2) Number of transcripts\cr
 (3) Original number of input transcripts
}
\description{
Finds compatible exon in annotation with the one present in roi object
}
\examples{
compatible_exons <- findCompatibleExon(valid_cds,roi)
}
\author{
Diana Low
}
