% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_splinter.R
\name{findExactOverlaps}
\alias{findExactOverlaps}
\title{findExactOverlaps}
\usage{
findExactOverlaps(query, subject, sequential = FALSE, verbose = FALSE)
}
\arguments{
\item{query}{GRanges object}

\item{subject}{GRanges object}

\item{sequential}{logical. TRUE if exons are sequential.}

\item{verbose}{logical. report intermediate output}
}
\value{
Hits object
}
\description{
Internal function similar to findSpliceOverlaps but only preserves internal
flanks
}
\author{
Diana Low
}
\keyword{internal}

