% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_splinter.R
\name{findTermination}
\alias{findTermination}
\title{findTermination}
\usage{
findTermination(s1)
}
\arguments{
\item{s1}{character. protein sequence}
}
\value{
list containing\cr
  (1) stop1 : stop position\cr
  (2) s1 : sequence truncated to first stop
}
\description{
Internal function to find the first stop codon that occurs in the AA sequence, returns their position
and the resulting truncated protein
}
\author{
Diana LOW
}
\keyword{internal}
