% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_splinter.R
\name{makeUniqueIDs}
\alias{makeUniqueIDs}
\title{makeUniqueIDs}
\usage{
makeUniqueIDs(ddata)
}
\arguments{
\item{ddata}{extractSpliceEvents object}
}
\value{
original extractSpliceEvents list object with unique ID
   appended to data accessor
}
\description{
Makes unique ID names from event location
}
\examples{
data_with_id<-makeUniqueIDs(splice_data)
}
\author{
Diana Low
}
