% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_splinter.R
\name{matchExons}
\alias{matchExons}
\title{matchExons}
\usage{
matchExons(ref, subject)
}
\arguments{
\item{ref}{GRanges object}

\item{subject}{GRanges object}
}
\value{
logical. check if exons match (TRUE) or not (FALSE)
}
\description{
Internal function to help match the inner coordinates of a 2/3 cassette
checks if reference and subject matches
}
\author{
Diana Low
}
\keyword{internal}

