% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_splinter.R
\name{plot_seqlogo}
\alias{plot_seqlogo}
\title{plotting sequence logo}
\usage{
plot_seqlogo(fasta_seq)
}
\arguments{
\item{fasta_seq}{DNAStringSet or path to fasta-formatted file}
}
\value{
sequence logo image
}
\description{
Plots the sequence logo of a given set of FASTA sequences
}
\examples{
head(splice_fasta)
plot_seqlogo(Biostrings::DNAStringSet(splice_fasta$V2))
}
\author{
Diana Low
}
