% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_splinter.R
\name{psiPlot}
\alias{psiPlot}
\title{psiPlot}
\usage{
psiPlot(df = NULL, type = "MATS", sample_labels = c("Sample 1",
  "Sample 2"))
}
\arguments{
\item{df}{data.frame containing PSI values}

\item{type}{character. either 'MATS' output (will read in MATS headers) or
'generic' (provide 4 or 6 column data.frame)}

\item{sample_labels}{x-axis labels for the plot}
}
\value{
bar plot of PSI values
}
\description{
Plots percentage spliced in (PSI) values in terms of inclusion levels
}
\examples{
#we give inclusion and skipped numbers as reads
#this will be converted into percentages
df<-data.frame(inclusion1=c("6,4,6"),skipped1=c("10,12,12"),inclusion2=c("15,15,15"),
     skipped2=c("3,3,4"),stringsAsFactors = FALSE)
psiPlot(df,type='generic')

}
\author{
Diana Low
}

