% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_splinter.R
\name{removeRegion}
\alias{removeRegion}
\title{removeRegion}
\usage{
removeRegion(subject, roi)
}
\arguments{
\item{subject}{GRanges or GrangesList object}

\item{roi}{\code{\link{makeROI}} object containing GRanges range (to remove)}
}
\value{
GRanges object
}
\description{
removes a region (exon) from a GRanges or GRangesList
}
\author{
Diana Low

# Removes the exon defined in roi GRanges object from a GRanges/GRangesList
compatible_cds$hits[[1]]
region_minus_exon<-removeRegion(compatible_cds$hits[[1]],roi)
}
