% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_splinter.R
\name{remvalue}
\alias{remvalue}
\title{remvalue}
\usage{
remvalue(x)
}
\arguments{
\item{x}{GRanges or GRangesList}
}
\value{
GRanges or GRangesList
}
\description{
helper function to remove metadata from GRanges object
used within metaremove
}
\keyword{internal}
