% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_splinter.R
\name{shapiroDensity}
\alias{shapiroDensity}
\title{shapiroDensity}
\usage{
shapiroDensity(ctrl_scores, treat_scores, sample = c(1, 2))
}
\arguments{
\item{ctrl_scores}{output of shapiroDonor or shapiroAcceptor}

\item{treat_scores}{output of shapiroDonor or shapiroAcceptor}

\item{sample}{samplenames}
}
\value{
density plot of Shapiro scores
}
\description{
convenience function for plotting Shapiro score density
}
\author{
Diana Low
}
