\name{splice_fasta}
\alias{splice_fasta}
\docType{data}
\title{
splice_fasta
}
\description{
splice_fasta
}
\usage{data("splice_fasta")}
\format{
  A data frame with 0 observations on the following 2 variables.
  \describe{
    \item{\code{V1}}{a numeric vector}
    \item{\code{V2}}{a numeric vector}
  }
}
\value{
Dataframe of region and fasta sequence
}
\examples{
data(splice_fasta)
}
\keyword{datasets}
