% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primerpcr.R
\name{splitPCRhit}
\alias{splitPCRhit}
\title{splitPCRhit}
\usage{
splitPCRhit(res, hitlist)
}
\arguments{
\item{res}{result from \code{\link{getPCRsizes}}}

\item{hitlist}{\code{\link{findCompatibleEvents}} object}
}
\value{
list of 2 data.frame objects with isoform name (ID) and
  length of PCR product (bp) matching Type 1 or Type 2 transcripts
}
\description{
splits the PCR alignment into the two AS conditions
}
\examples{
## as getPCRsizes gives you all PCR bands when the primers are used,
## splitPCRhit will determine which bands are relevant to the target
relevant_pcr_bands<-splitPCRhit(pcr_result1,compatible_tx)
}
\author{
Diana Low
}
