% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_splinter.R
\name{extractSpliceSites}
\alias{extractSpliceSites}
\title{extractSpliceSites}
\usage{
extractSpliceSites(file, splicetype = "SE", site = "donor", fdr = 1,
  motif_range = c(-3, 6), inclusion = 0)
}
\arguments{
\item{file}{character MATS \url{http://rnaseq-mats.sourceforge.net/} output filename
or data.frame output from extractSpliceEvents}

\item{splicetype}{character either SE (skipped exon) or RI (retained intron)}

\item{site}{character donor or acceptor}

\item{fdr}{numeric false discovery rate filter range [0,1]}

\item{motif_range}{numeric vector of splice position to extract}

\item{inclusion}{numeric fraction,takes absolute value}
}
\value{
GRanges object
}
\description{
Extracts and formats to bed the location of target, upstream and downstream
splice sites
}
\examples{
data_path<-system.file("extdata",package="SPLINTER")
splice_data<-extractSpliceEvents(data=paste(data_path,"/skipped_exons.txt",sep=""))
splice_sites<-extractSpliceSites(splice_data$data)

## or
#splice_sites<-extractSpliceSites(file=paste(data_path,"/skipped_exons.txt",sep=""))
}
\author{
Diana Low
}
\seealso{
\url{http://rnaseq-mats.sourceforge.net/user_guide.htm} for MATS file definition
}

