% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_splinter.R
\name{makeROI}
\alias{makeROI}
\title{makeROI}
\usage{
makeROI(info, itemnum = 1)
}
\arguments{
\item{info}{\code{\link{extractSpliceEvents}} object}

\item{itemnum}{integer. row number of item in \code{info}}
}
\value{
a list containing\cr
  (1) type : splice type\cr
  (2) name : ID of transcript\cr
  (3) roi  : GRanges object of splice site\cr
  (4) flank : GRanges object of flanking exons of splice site\cr
  (5) roi_range : GRangesList of splice site and it's alternative outcome based on \code{type}
}
\description{
Creates an object to store information about the splice site (region of interest)
including flanking regions and alternative splice outcome
}
\examples{
roi <- makeROI(splice_data,1)
}
\author{
Diana Low
}
